pyproject.toml
[project]
name = "emote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]
